BaseEntity, BaseIdEntity 분리 처리 BaseIdEntity 클래스에 공통 식별자 정수형으로 정의 비즈니스 도메인, 테이블 스키마 변경에 따른 별도의 DB 및 profile 정의

+++ src/main/java/com/koke/koke_backend/celebrity/repository/CelebrityCommentRepository.java
@@ -3,5 +3,5 @@ package com.koke.koke_backend.celebrity.repository;
 import com.koke.koke_backend.celebrity.entity.CelebrityComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CelebrityCommentRepository extends JpaRepository<CelebrityComment, Integer> {
+public interface CelebrityCommentRepository extends JpaRepository<CelebrityComment, Long> {
 }
